Updated the IDV examples to add routes and use case controllers to handle different IDV session use cases.
