+++ examples/idv/src/controllers/index.js
@@ -1,5 +1,5 @@
 const indexController = require('./index.controller');
-const indexControllerWithFaceComparison = require('./index.controller.with.face.comparison');
+const useCasesControllers = require('./use-cases');
 const successController = require('./success.controller');
 const configController = require('./config.controller');
 const mediaController = require('./media.controller');
@@ -7,7 +7,7 @@ const errorController = require('./error.controller');
 
 module.exports = {
   indexController,
-  indexControllerWithFaceComparison,
+  useCasesControllers,
   successController,
   configController,
   mediaController,
