+++ examples/idv/src/controllers/use-cases/face.match.check.controller.js
@@ -1,0 +1,76 @@
+const {
+  IDVClient,
+  SessionSpecificationBuilder,
+  RequestedLivenessCheckBuilder,
+  RequestedFaceMatchCheckBuilder,
+  SdkConfigBuilder,
+  RequiredIdDocumentBuilder,
+} = require('yoti');
+const config = require('../../../config');
+
+/**
+ * Create an IDV session.
+ */
+async function createSession() {
+  const idvClient = new IDVClient(
+    config.YOTI_CLIENT_SDK_ID,
+    config.YOTI_PEM
+  );
+
+  const sessionSpec = new SessionSpecificationBuilder()
+    .withClientSessionTokenTtl(600)
+    .withResourcesTtl(90000)
+    .withUserTrackingId('some-user-tracking-id')
+    // For zoom liveness check [only one type of liveness check to be enabled at a time]
+    .withRequestedCheck(
+      new RequestedLivenessCheckBuilder()
+        .forZoomLiveness()
+        .build()
+    )
+    // For static liveness check
+    // .withRequestedCheck(
+    //   new RequestedLivenessCheckBuilder()
+    //     .forStaticLiveness()
+    //     .build()
+    // )
+    .withRequestedCheck(
+      new RequestedFaceMatchCheckBuilder()
+        .withManualCheckNever()
+        .build()
+    )
+    .withRequiredDocument(
+      (new RequiredIdDocumentBuilder()).build()
+    )
+    .withSdkConfig(
+      new SdkConfigBuilder()
+        .withAllowsCameraAndUpload()
+        .withLocale('en-GB')
+        .withPresetIssuingCountry('GBR')
+        .withSuccessUrl(`${config.YOTI_APP_BASE_URL}/success`)
+        .withErrorUrl(`${config.YOTI_APP_BASE_URL}/error`)
+        .withPrivacyPolicyUrl(`${config.YOTI_APP_BASE_URL}/privacy-policy`)
+        .withEarlyBiometricConsentFlow() // or withJustInTimeBiometricConsentFlow()
+        .withIdDocumentTextExtractionGenericRetries(5)
+        .withIdDocumentTextExtractionReclassificationRetries(5)
+        .build()
+    )
+    .build();
+
+  return idvClient.createSession(sessionSpec);
+}
+
+module.exports = async (req, res) => {
+  try {
+    const session = await createSession();
+
+    req.session.IDV_SESSION_ID = session.getSessionId();
+    req.session.IDV_SESSION_TOKEN = session.getClientSessionToken();
+
+    res.render('pages/session', {
+      sessionId: req.session.IDV_SESSION_ID,
+      iframeUrl: `${config.YOTI_IDV_IFRAME_URL}?sessionID=${req.session.IDV_SESSION_ID}&sessionToken=${req.session.IDV_SESSION_TOKEN}`,
+    });
+  } catch (error) {
+    res.render('pages/error', { error });
+  }
+};
