+++ examples/idv/src/controllers/use-cases/index.js
@@ -1,0 +1,13 @@
+const authenticityAndIdentityCheckController = require('./authenticity.and.identity.check.controller');
+const documentComparisonCheckController = require('./document.comparison.check.controller');
+const faceComparisonCheckController = require('./face.comparison.check.controller');
+const faceMatchCheckController = require('./face.match.check.controller');
+const watchlistCheckController = require('./watchlist.check.controller');
+
+module.exports = {
+  authenticityAndIdentityCheckController,
+  documentComparisonCheckController,
+  faceComparisonCheckController,
+  faceMatchCheckController,
+  watchlistCheckController,
+};
