+++ examples/idv/src/useCasesList.js
@@ -1,0 +1,29 @@
+const useCasesList = [
+  {
+    id: 'documentAuthenticityAndIdentity',
+    name: 'Document authenticity and Identity check',
+    href: '/use-cases/document-authenticity-and-identity-check',
+  },
+  {
+    id: 'documentComparison',
+    name: 'Document comparison check',
+    href: '/use-cases/document-comparison-check',
+  },
+  {
+    id: 'faceComparison',
+    name: 'Face comparison check',
+    href: '/use-cases/face-comparison-check',
+  },
+  {
+    id: 'faceMatch',
+    name: 'Face match check',
+    href: '/use-cases/face-match-check',
+  },
+  {
+    id: 'watchlist',
+    name: 'Watchlist check',
+    href: '/use-cases/watchlist-check',
+  },
+];
+
+module.exports = useCasesList;
